Split file-backed configuration module
